requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
numpy<2
Pillow
PyYAML
torch
torchvision
